requirements.txt
opencv-python-headless
pdf2image
PyPDF2
python-dotenv
azure-identity
azure-storage-blob
azure-functions
requests